added apartmentsBySaleStatus stats added apartmentsPerFloorPerHouse stats changed floorNumber to Integer

+++ src/main/java/kattsyn/dev/ApartmentsUnderConstruction/dtos/FloorDTO.java
@@ -11,7 +11,7 @@ import lombok.Setter;
 @Setter
 public class FloorDTO {
 
-    private Byte floorNumber;
+    private Integer floorNumber;
     private String floorPlan;
     private Long houseId;
